raise error in base; impl in subclass
